+++ src/components/QRCode/QRCode.jsx
