+++ src/components/TrafficLight/TrafficLight.jsx
